Revert "Refactor: passport 로직 제거 및 domains 폴더 내부로 비즈니스 로직(auth, user, social-auth) 모듈 이동 #11"
This reverts commit 7b024ddf163d034cb2a0c83b1456829ab80e778a.
