+++ src/app.module.ts
@@ -5,9 +5,8 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { OrmConfig } from './common/database/orm-config';
 import { DataSource, DataSourceOptions } from 'typeorm';
-import { UserModule } from './domains/user/user.module';
-import { AuthModule } from './domains/auth/auth.module';
-import { SocialAuthModule } from './domains/social-auth/social-auth.module';
+import { UserModule } from './user/user.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
@@ -19,11 +18,13 @@ import { SocialAuthModule } from './domains/social-auth/social-auth.module';
     }),
     ConfigModule.forRoot({
       isGlobal: true,
-      envFilePath: process.env.NODE_ENV === 'development' ? '.env.development' : '.env.production',
+      envFilePath:
+        process.env.NODE_ENV === 'development'
+          ? '.env.development'
+          : '.env.production',
     }),
     UserModule,
     AuthModule,
-    SocialAuthModule,
   ],
   controllers: [AppController],
   providers: [AppService],
