+++ src/auth/auth.module.ts
@@ -1,14 +1,15 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { UserModule } from 'src/domains/user/user.module';
+import { UserModule } from 'src/user/user.module';
 import { JwtModule } from '@nestjs/jwt';
-import { SocialAuthModule } from '../social-auth/social-auth.module';
+import { LocalStrategy } from './strategy/local.strategy';
+import { JwtStrategy } from './strategy/jwt.strategy';
 
 @Module({
-  imports: [JwtModule.register({}), UserModule, SocialAuthModule],
+  imports: [JwtModule.register({}), UserModule],
   controllers: [AuthController],
-  providers: [AuthService],
+  providers: [AuthService, LocalStrategy, JwtStrategy],
   exports: [AuthService],
 })
 export class AuthModule {}
