--- src/domains/auth/consts/social-type.const.ts
@@ -1,6 +1,0 @@
-export enum SocialUserType {
-  GOOGLE = 'google',
-  APPLE = 'apple',
-  NAVER = 'naver',
-  KAKAO = 'kakao',
-}
