--- src/domains/auth/consts/token-type.const.ts
@@ -1,7 +1,0 @@
-/** Token Type */
-export const AuthTokenType = Object.freeze({
-  ACCESS_TOKEN: 'at',
-  REFRESH_TOKEN: 'rt',
-} as const);
-
-type AuthTokenType = (typeof AuthTokenType)[keyof typeof AuthTokenType];
