--- src/domains/social-auth/social-auth.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SocialAuthService } from './social-auth.service';
-
-@Module({
-  providers: [SocialAuthService],
-  exports: [SocialAuthService],
-})
-export class SocialAuthModule {}
